Move LookupFuture to resolver module
